feat(cash-gift): separated payment schemas and routes from gift
